1. Модульные тесты. [#378848]

+++ src/ru/job4j/converter/Converter.java
@@ -4,21 +4,21 @@ import java.sql.SQLOutput;
 
 public class Converter {
 
-    public static double rubleToEuro(int value) {
-        double rsl = value / 70;
+    public static int rubleToEuro(int value) {
+        int rsl = value / 70;
         return rsl;
     }
 
-    public static double rubleToDollar(int value) {
-        double rsl = value / 60;
+    public static int rubleToDollar(int value) {
+        int rsl = value / 60;
         return rsl;
     }
 
     public static void main(String[] args) {
         int in = 140;
-        double expected = 2;
-        double outEuro = Converter.rubleToEuro(in);
-        double outDollar = Converter.rubleToDollar(in);
+        int expected = 2;
+        int outEuro = Converter.rubleToEuro(in);
+        int outDollar = Converter.rubleToDollar(in);
         System.out.println("140 rubles are " + outEuro + " euro. Test result : " + (outEuro == expected));
         System.out.println("140 rubles are " + outDollar + " dollar. Test result : " + (outDollar == expected));
     }
